Create shortcut Bus from BusObject
