+++ Assets/Unibus/BusObject.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0a286885f1c0049ca907a9c84337f0b7
+timeCreated: 1465739621
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
